// src/web-ui/src/app/models/viewModels.ts
import { responseStatus, messageType} from '../models/enums';

export interface IEnvironment {
		env: string;
}

export interface IAppConfig {
		env: string;
		loginUrl: string;
		
}

 
export interface IResponse<T> {
		data: T;
		status: responseStatus;
		messageKey: string;
		apiUrl: string;
}

export interface IAPIResponse<T> {
		data?: T;
		code: string;
		message: string;
} 

export interface ILoginDetail {
		id: number;
		userId: number;
		userName: string;
		accessToken: string;
		expires: Date;
		loggedinDate?: Date;
		isLoggedIn?: boolean;
}


export interface ILoginUser {
		password: string;
		userName: string;
}
 
export interface IKeyValue {
		id: number;
		name: string;
} 
export interface IKeyData {
		key: string;
		data: any;
}
 